Extract TestHttp into its own file
This better reflects the structure in *src/main* and removes some noise
from *MessagesQueuesSpec.scala*.
